requirements.txt
paho-mqtt
python-etcd
serial